Support multiple compensations bound to one key

+++ src/index.es6
@@ -3,11 +3,18 @@ import request from 'request-promise';
 
 module.exports = (options = {}) => {
   let client = redis.createClient(options);
-  return {
-    client,
-    add: (id, requestOptions = {}) => {
-      const key = id.toString();
-      return new Promise((resolve, reject) => {
+
+  function add(key, serviceKey = null, requestOptions = {}) {
+    return new Promise((resolve, reject) => {
+      if (serviceKey) {
+        client.hset(key, serviceKey, JSON.stringify(requestOptions), (err, res) => {
+          if (!err) {
+            resolve(res);
+          } else {
+            reject(err);
+          }
+        });
+      } else {
         client.set(key, JSON.stringify(requestOptions), (err, res) => {
           if (!err) {
             resolve(res);
@@ -15,11 +22,21 @@ module.exports = (options = {}) => {
             reject(err);
           }
         });
-      });
-    },
-    remove: id => {
-      const key = id.toString();
-      return new Promise((resolve, reject) => {
+      }
+    });
+  }
+
+  function remove(key, serviceKey = null) {
+    return new Promise((resolve, reject) => {
+      if (serviceKey) {
+        client.hdel(key, serviceKey, (err, res) => {
+          if (!err) {
+            resolve(res);
+          } else {
+            reject(err);
+          }
+        });
+      } else {
         client.del(key, (err, res) => {
           if (!err) {
             resolve(res);
@@ -27,23 +44,67 @@ module.exports = (options = {}) => {
             reject(err);
           }
         });
-      });
-    },
-    run: id => {
-      const key = id.toString();
-      return new Promise((resolve, reject) => {
-        client.get(key, (err, data) => {
+      }
+    });
+  }
+
+  function run(key, serviceKey = null) {
+    return new Promise((resolve, reject) => {
+      if (serviceKey) {
+        client.hget(key, serviceKey, (err, data) => {
           if (!err) {
-            const requestOptions = JSON.parse(data);
             // Call compensating action request
-            request(requestOptions)
+            request(JSON.parse(compensation))
               .then(res => resolve(res))
               .catch(err => reject(err));
           } else {
             reject(err);
           }
         });
+      } else {
+        client.get(key, (err, compensation) => {
+          if (!err) {
+            // Call compensating action request
+            request(JSON.parse(compensation))
+              .then(res => {
+                // Remove key on compensation success
+                remove(key)
+                  // Return result of compensation
+                  .then(_res => {
+                    resolve(res);
+                  });
+              })
+              .catch(err => reject(err));
+          } else {
+            reject(err);
+          }
+        });
+      }
+    });
+  }
+
+  function runAll(key) {
+    return new Promise((resolve, reject) => {
+      client.hgetall(key, (err, data) => {
+        if (!err) {
+          let promises = [];
+          // Call compensating action requests
+          data.forEach(compensation => {
+            promises.push(request(JSON.parse(compensation)));
+          });
+          resolve(Promise.all(promises));
+        } else {
+          reject(err);
+        }
       });
-    }
+    });
+  }
+
+  return {
+    client,
+    add,
+    remove,
+    run,
+    runAll
   }
 };
